backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
supabase==2.11.0
pydantic-settings>=2.7.0
python-dotenv>=1.1.0
python-jose[cryptography]>=3.4.0
yfinance>=0.2.40
jugaad-data>=0.27
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.11.0
requests>=2.31.0
beautifulsoup4>=4.12.0
